+导出SQL文件功能

+++ src/test/java/com/excel/database/converter/service/MysqlDatabaseServiceTest.java
@@ -26,6 +26,12 @@ public class MysqlDatabaseServiceTest {
     @Qualifier("Mysql")
     private IDatabaseService databaseService;
 
+    @Resource
+    @Qualifier("Mysql")
+    private IDatabaseExportService databaseExportService;
+
+    private String sqlFilePath = System.getProperty("user.dir") + File.separator + "mysqlTest.sql";
+
     private String databaseName = "mysqlTest";
 
     private String tableName = "tableTest";
@@ -71,6 +77,35 @@ public class MysqlDatabaseServiceTest {
 
             boolean result = databaseService.insert(databaseUrl, tableName, fields, data);
             System.out.println(String.format("插入数据表%s。", result ? "成功" : "失败"));
+
+            if (result) {
+                result = databaseService.dropTable(databaseUrl, tableName);
+                System.out.println(String.format("删除数据表%s。", result ? "成功" : "失败"));
+            }
+        }
+    }
+
+    @Test
+    public void export() throws Exception {
+        String databaseUrl = databaseService.createDatabase(databaseName);
+        System.out.println(databaseUrl);
+
+        Map<Integer, String> fields = new HashMap<>(4);
+        fields.put(1, "字段1");
+        fields.put(2, "字段2");
+        fields.put(3, "字段3");
+
+        if (databaseService.crateTable(databaseUrl, tableName, fields)) {
+            Map<Integer, String> data = new HashMap<>(4);
+            data.put(0, "数据1");
+            data.put(1, "数据2");
+            data.put(2, "数据3");
+
+            boolean result = databaseService.insert(databaseUrl, tableName, fields, data);
+            System.out.println(String.format("插入数据表%s。", result ? "成功" : "失败"));
+
+            result = databaseExportService.exportDatabaseStructureAndData(databaseUrl, databaseName, sqlFilePath);
+            System.out.println(String.format("导出数据表%s。", result ? "成功" : "失败"));
 
             if (result) {
                 result = databaseService.dropTable(databaseUrl, tableName);
